requirements.txt
absl-py==1.3.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.3
aiosignal==1.2.0
ale-py==0.8.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
asttokens==3.0.0
async-timeout==4.0.2
asyncio==3.4.3
asynctest==0.13.0
attrs==25.3.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
blessed==1.25.0
blinker==1.8.2
braceexpand==0.1.7
cachetools==5.2.0
ccimport==0.4.4
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==2.1.1
check-wheel-contents==0.6.1
click==8.1.3
clip @ file:///home/runling/workplace/CLIP
cloudpickle==1.6.0
colorama==0.4.6
comm==0.2.2
commonmark==0.9.1
ConfigArgParse==1.5.3
configparser==7.1.0
contourpy==1.1.1
crcmod==1.7
cryptography==45.0.2
cumm==0.4.11
cycler==0.11.0
Cython==0.29.33
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
disutils==1.4.32.post2
dm-tree==0.1.8
docker-pycreds==0.4.0
docutils==0.20.1
e==1.4.5
einops==0.8.1
etils==1.3.0
eval_type_backport==0.2.2
exceptiongroup==1.3.1
executing==2.1.0
Farama-Notifications==0.0.4
fastapi==0.124.4
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.14.0
fire==0.7.0
Flask==2.2.2
fonttools==4.38.0
freetype-py==2.5.1
frozenlist==1.3.1
fsspec==2024.12.0
ftfy==6.2.3
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.9
GitPython==3.1.45
glfw==2.10.0
google-auth==2.43.0
google-auth-oauthlib==1.0.0
gpustat==1.1.1
grpcio==1.50.0
gym==0.23.1
gym-notices==0.0.8
gymnasium==0.28.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
id==1.5.0
idna==3.4
imageio==2.22.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.13.0
importlib_resources==6.4.5
iopath==0.1.10
ipdb==0.13.13
ipython==8.12.3
ipywidgets==8.1.5
isaacgym @ file:///home/runling/workplace/isaacgym/python
-e git+https://github.com/isaac-sim/IsaacGymEnvs.git@aeed298638a1f7b5421b38f5f3cc2d1079b6d9c3#egg=isaacgymenvs
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax-jumpy==1.0.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.17.3
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keyring==25.5.0
kiwisolver==1.4.4
kornia==0.6.8
lark==1.2.2
legged-gym @ file:///home/runling/workplace/city_isaacgym_ours
lpips==0.1.4
lxml==5.3.1
lz4==4.3.3
Mako==1.3.8
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmcls==0.25.0
mmcv-full==1.7.2
mmengine==0.10.7
mmsegmentation==0.30.0
model-index==0.1.11
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.1
mujoco==3.2.3
multidict==6.0.2
narwhals==1.42.1
nbformat==5.7.0
nest-asyncio==1.6.0
networkx==2.2
nh3==0.2.21
ninja==1.11.1
numpy==1.23.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.17.0
open3d==0.17.0
open_clip_torch==2.32.0
opencv-python==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.3
ordered-set==4.1.0
orjson==3.10.15
oss2==2.17.0
packaging==24.2
pandas==1.3.5
parso==0.8.4
pathtools==0.1.2
pccm==0.4.16
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
plyfile==0.7.4
portalocker==3.0.0
prettytable==3.11.0
promise==2.3
prompt_toolkit==3.0.50
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.1
pycollada==0.6
pycparser==2.22
pycryptodome==3.23.0
pycuda==2024.1.2
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.1.0
pyglet==1.5.0
Pygments==2.13.0
PyOpenGL==3.1.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrender==0.1.45
pyrsistent==0.20.0
pysdf==0.1.9
python-dateutil==2.8.2
python-multipart==0.0.20
pytools==2024.1.14
pytorch3d==0.7.8
pytz==2023.4
PyVirtualDisplay==3.0
PyWavelets==1.3.0
PyYAML==6.0
ray==2.10.0
readme_renderer==43.0
referencing==0.35.1
regex==2024.11.6
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rerun-sdk==0.19.0
retrying==1.3.4
rfc3986==2.0.0
rich==13.4.2
rl_games @ file:///home/runling/workplace/city_isaacgym_ours/rl_games
rpds-py==0.20.1
rsa==4.9
rsl-rl-lib==2.3.3
ruff==0.14.9
safetensors==0.5.3
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.8.0
seaborn==0.13.2
SecretStorage==3.3.3
semantic-version==2.10.0
sentry-sdk==2.46.0
setproctitle==1.3.2
shellingham==1.5.4
Shimmy==0.2.1
shortuuid==1.0.11
six==1.16.0
skrl==1.4.3
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.6
spconv==2.3.6
stack-data==0.6.3
starlette==0.44.0
subprocess32==3.5.4
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2021.11.2
timm==1.0.22
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.12.0
torch @ file:///home/runling/workplace/torch-2.1.0%2Bcu121-cp38-cp38-linux_x86_64.whl#sha256=4c83190ad649c77adaf6e1c616998f10598db696912ea7a410831632890b49bf
torch-scatter==2.1.2
torchvision @ file:///home/runling/workplace/torchvision-0.16.0%2Bcu121-cp38-cp38-linux_x86_64.whl#sha256=967c0f8ada2abeed430b33e83ce7b3de5233675c8b020c18c1255dca065b4add
tqdm==4.65.2
traitlets==5.14.3
transformers==4.46.3
trimesh==3.23.5
triton==2.1.0
twine==6.1.0
typer==0.20.0
typing_extensions==4.12.2
tzdata==2024.2
urdfpy==0.0.22
urllib3==1.26.20
uvicorn==0.33.0
wandb==0.12.11
warp-lang==0.10.1
wcwidth==0.2.13
webdataset==0.2.100
websockets==12.0
Werkzeug==2.2.3
wheel-filename==1.4.2
widgetsnbextension==4.0.13
yacs==0.1.8
yapf==0.30.0
yarl==1.8.1
yaspin==2.5.0
zipp==3.10.0
